restructure test files (still only running ios)
